OrderFilled fixes; updated migrations
